Remove interface/impl separation in quality model code
