+++ analyzer/src/main/java/eubrazil/atmosphere/qualitymodel/Attribute.java
@@ -2,80 +2,56 @@
  */
 package eubrazil.atmosphere.qualitymodel;
 
-
 /**
- * <!-- begin-user-doc -->
- * A representation of the model object '<em><b>Attribute</b></em>'.
- * <!-- end-user-doc -->
- *
- * <p>
- * The following features are supported:
- * </p>
- * <ul>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.Attribute#getName <em>Name</em>}</li>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.Attribute#getCurrentValue <em>Current Value</em>}</li>
- * </ul>
- *
- * @model abstract="true"
  * @generated
  */
-public interface Attribute {
+public abstract class Attribute {
 	/**
-	 * Returns the value of the '<em><b>Name</b></em>' attribute.
-	 * <!-- begin-user-doc -->
-	 * <p>
-	 * If the meaning of the '<em>Name</em>' attribute isn't clear,
-	 * there really should be more of a description here...
-	 * </p>
-	 * <!-- end-user-doc -->
-	 * @return the value of the '<em>Name</em>' attribute.
-	 * @see #setName(String)
-	 * @model id="true" required="true"
-	 * @generated
+	 * @ordered
 	 */
-	String getName();
+	protected static final String NAME_EDEFAULT = null;
 
 	/**
-	 * Sets the value of the '{@link eubrazil.atmosphere.qualitymodel.Attribute#getName <em>Name</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @param value the new value of the '<em>Name</em>' attribute.
-	 * @see #getName()
 	 * @generated
 	 */
-	void setName(String value);
+	protected String name = NAME_EDEFAULT;
 
 	/**
-	 * Returns the value of the '<em><b>Current Value</b></em>' attribute.
-	 * <!-- begin-user-doc -->
-	 * <p>
-	 * If the meaning of the '<em>Current Value</em>' attribute isn't clear,
-	 * there really should be more of a description here...
-	 * </p>
-	 * <!-- end-user-doc -->
-	 * @return the value of the '<em>Current Value</em>' attribute.
-	 * @see #setCurrentValue(double)
-	 * @model required="true"
 	 * @generated
 	 */
-	double getCurrentValue();
+	public Attribute() {
+		super();
+	}
 
 	/**
-	 * Sets the value of the '{@link eubrazil.atmosphere.qualitymodel.Attribute#getCurrentValue <em>Current Value</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @param value the new value of the '<em>Current Value</em>' attribute.
-	 * @see #getCurrentValue()
 	 * @generated
 	 */
-	void setCurrentValue(double value);
+	public String getName() {
+		return name;
+	}
 
 	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @model required="true"
 	 * @generated
 	 */
-	double calculate(UserProfile user);
+	public void setName(String newName) {
+		name = newName;
+	}
+
+	/**
+	 * @generated NOT
+	 */
+	public abstract HistoricalData calculate(UserProfile user);
+
+	/**
+	 * @generated
+	 */
+	@Override
+	public String toString() {
+		StringBuffer result = new StringBuffer(super.toString());
+		result.append(" (name: ");
+		result.append(name);
+		result.append(')');
+		return result.toString();
+	}
 
 } // Attribute
