+++ analyzer/src/main/java/eubrazil/atmosphere/qualitymodel/CompositeAttribute.java
@@ -2,66 +2,120 @@
  */
 package eubrazil.atmosphere.qualitymodel;
 
+import java.sql.Timestamp;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
- * <!-- begin-user-doc -->
- * A representation of the model object '<em><b>Composite Attribute</b></em>'.
- * <!-- end-user-doc -->
- *
- * <p>
- * The following features are supported:
- * </p>
- * <ul>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.CompositeAttribute#getChildren <em>Children</em>}</li>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.CompositeAttribute#getOperator <em>Operator</em>}</li>
- * </ul>
- *
- * @model
  * @generated
  */
-public interface CompositeAttribute extends Attribute {
+public class CompositeAttribute extends Attribute {
 	/**
-	 * Returns the value of the '<em><b>Children</b></em>' containment reference list.
-	 * The list contents are of type {@link eubrazil.atmosphere.qualitymodel.Attribute}.
-	 * <!-- begin-user-doc -->
-	 * <p>
-	 * If the meaning of the '<em>Children</em>' containment reference list isn't clear,
-	 * there really should be more of a description here...
-	 * </p>
-	 * <!-- end-user-doc -->
-	 * @return the value of the '<em>Children</em>' containment reference list.
-	 * @model containment="true" required="true"
 	 * @generated
 	 */
-	List<Attribute> getChildren();
+	protected List<Attribute> children;
 
 	/**
-	 * Returns the value of the '<em><b>Operator</b></em>' attribute.
-	 * The literals are from the enumeration {@link eubrazil.atmosphere.qualitymodel.AttributeAggregationOperator}.
-	 * <!-- begin-user-doc -->
-	 * <p>
-	 * If the meaning of the '<em>Operator</em>' attribute isn't clear,
-	 * there really should be more of a description here...
-	 * </p>
-	 * <!-- end-user-doc -->
-	 * @return the value of the '<em>Operator</em>' attribute.
-	 * @see eubrazil.atmosphere.qualitymodel.AttributeAggregationOperator
-	 * @see #setOperator(AttributeAggregationOperator)
-	 * @model required="true"
 	 * @generated
 	 */
-	AttributeAggregationOperator getOperator();
+	protected static final AttributeAggregationOperator OPERATOR_EDEFAULT = AttributeAggregationOperator.NEUTRALITY;
 
 	/**
-	 * Sets the value of the '{@link eubrazil.atmosphere.qualitymodel.CompositeAttribute#getOperator <em>Operator</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @param value the new value of the '<em>Operator</em>' attribute.
-	 * @see eubrazil.atmosphere.qualitymodel.AttributeAggregationOperator
-	 * @see #getOperator()
 	 * @generated
 	 */
-	void setOperator(AttributeAggregationOperator value);
+	protected AttributeAggregationOperator operator = OPERATOR_EDEFAULT;
+
+	/**
+	 * @generated
+	 */
+	public CompositeAttribute() {
+		super();
+	}
+
+	/**
+	 * @generated
+	 */
+	public List<Attribute> getChildren() {
+		if (children == null) {
+			children = new ArrayList<Attribute>();
+		}
+		return children;
+	}
+
+	/**
+	 * @generated
+	 */
+	public AttributeAggregationOperator getOperator() {
+		return operator;
+	}
+
+	/**
+	 * @generated
+	 */
+	public void setOperator(AttributeAggregationOperator newOperator) {
+		operator = newOperator == null ? OPERATOR_EDEFAULT : newOperator;
+	}
+
+	/**
+	 * @generated
+	 */
+	protected double calculateNeutrality() {
+		// TODO: implement this method
+		// Ensure that you remove @generated or mark it @generated NOT
+		throw new UnsupportedOperationException();
+	}
+
+	/**
+	 * @generated
+	 */
+	protected double calculateSimultaneity() {
+		// TODO: implement this method
+		// Ensure that you remove @generated or mark it @generated NOT
+		throw new UnsupportedOperationException();
+	}
+
+	/**
+	 * @generated
+	 */
+	protected double calculateReplaceability() {
+		// TODO: implement this method
+		// Ensure that you remove @generated or mark it @generated NOT
+		throw new UnsupportedOperationException();
+	}
+
+	/**
+	 * @generated
+	 */
+	@Override
+	public String toString() {
+		StringBuffer result = new StringBuffer(super.toString());
+		result.append(" (operator: ");
+		result.append(operator);
+		result.append(')');
+		return result.toString();
+	}
+	
+	@Override
+	public HistoricalData calculate(UserProfile user) {
+		HistoricalData d = new HistoricalData();
+		d.setInstant(new Timestamp(System.currentTimeMillis()));
+		d.attribute = this;
+		
+		switch (operator) {
+		case NEUTRALITY:
+			d.value = calculateNeutrality();
+			break;
+		case REPLACEABILITY:
+			d.value = calculateReplaceability();
+			break;
+		case SIMULTANEITY:
+			d.value = calculateSimultaneity();
+			break;
+		default:
+			throw new UnsupportedOperationException();
+		}
+		
+		return d;
+	}
 
 } // CompositeAttribute
