--- analyzer/src/main/java/eubrazil/atmosphere/qualitymodel/impl/AttributeImpl.java
@@ -1,135 +1,0 @@
-/**
- */
-package eubrazil.atmosphere.qualitymodel.impl;
-
-import eubrazil.atmosphere.qualitymodel.Attribute;
-import eubrazil.atmosphere.qualitymodel.UserProfile;
-
-/**
- * <!-- begin-user-doc -->
- * An implementation of the model object '<em><b>Attribute</b></em>'.
- * <!-- end-user-doc -->
- * <p>
- * The following features are implemented:
- * </p>
- * <ul>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.impl.AttributeImpl#getName <em>Name</em>}</li>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.impl.AttributeImpl#getCurrentValue <em>Current Value</em>}</li>
- * </ul>
- *
- * @generated
- */
-public abstract class AttributeImpl implements Attribute {
-	/**
-	 * The default value of the '{@link #getName() <em>Name</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getName()
-	 * @generated
-	 * @ordered
-	 */
-	protected static final String NAME_EDEFAULT = null;
-
-	/**
-	 * The cached value of the '{@link #getName() <em>Name</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getName()
-	 * @generated
-	 * @ordered
-	 */
-	protected String name = NAME_EDEFAULT;
-
-	/**
-	 * The default value of the '{@link #getCurrentValue() <em>Current Value</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getCurrentValue()
-	 * @generated
-	 * @ordered
-	 */
-	protected static final double CURRENT_VALUE_EDEFAULT = 0.0;
-
-	/**
-	 * The cached value of the '{@link #getCurrentValue() <em>Current Value</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getCurrentValue()
-	 * @generated
-	 * @ordered
-	 */
-	protected double currentValue = CURRENT_VALUE_EDEFAULT;
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	protected AttributeImpl() {
-		super();
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public String getName() {
-		return name;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public void setName(String newName) {
-		name = newName;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public double getCurrentValue() {
-		return currentValue;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public void setCurrentValue(double newCurrentValue) {
-		currentValue = newCurrentValue;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public double calculate(UserProfile user) {
-		// TODO: implement this method
-		// Ensure that you remove @generated or mark it @generated NOT
-		throw new UnsupportedOperationException();
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	@Override
-	public String toString() {
-		StringBuffer result = new StringBuffer(super.toString());
-		result.append(" (name: ");
-		result.append(name);
-		result.append(", currentValue: ");
-		result.append(currentValue);
-		result.append(')');
-		return result.toString();
-	}
-
-} //AttributeImpl
