--- analyzer/src/main/java/eubrazil/atmosphere/qualitymodel/impl/HistoricalDataImpl.java
@@ -1,155 +1,0 @@
-/**
- */
-package eubrazil.atmosphere.qualitymodel.impl;
-
-import eubrazil.atmosphere.qualitymodel.Attribute;
-import eubrazil.atmosphere.qualitymodel.HistoricalData;
-
-import java.sql.Timestamp;
-
-/**
- * <!-- begin-user-doc -->
- * An implementation of the model object '<em><b>Historical Data</b></em>'.
- * <!-- end-user-doc -->
- * <p>
- * The following features are implemented:
- * </p>
- * <ul>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.impl.HistoricalDataImpl#getInstant <em>Instant</em>}</li>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.impl.HistoricalDataImpl#getValue <em>Value</em>}</li>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.impl.HistoricalDataImpl#getAttribute <em>Attribute</em>}</li>
- * </ul>
- *
- * @generated
- */
-public class HistoricalDataImpl implements HistoricalData {
-	/**
-	 * The default value of the '{@link #getInstant() <em>Instant</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getInstant()
-	 * @generated
-	 * @ordered
-	 */
-	protected static final Timestamp INSTANT_EDEFAULT = null;
-
-	/**
-	 * The cached value of the '{@link #getInstant() <em>Instant</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getInstant()
-	 * @generated
-	 * @ordered
-	 */
-	protected Timestamp instant = INSTANT_EDEFAULT;
-
-	/**
-	 * The default value of the '{@link #getValue() <em>Value</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getValue()
-	 * @generated
-	 * @ordered
-	 */
-	protected static final double VALUE_EDEFAULT = 0.0;
-
-	/**
-	 * The cached value of the '{@link #getValue() <em>Value</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getValue()
-	 * @generated
-	 * @ordered
-	 */
-	protected double value = VALUE_EDEFAULT;
-
-	/**
-	 * The cached value of the '{@link #getAttribute() <em>Attribute</em>}' reference.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getAttribute()
-	 * @generated
-	 * @ordered
-	 */
-	protected Attribute attribute;
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	protected HistoricalDataImpl() {
-		super();
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public Timestamp getInstant() {
-		return instant;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public void setInstant(Timestamp newInstant) {
-		instant = newInstant;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public double getValue() {
-		return value;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public void setValue(double newValue) {
-		value = newValue;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public Attribute getAttribute() {
-		return attribute;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public void setAttribute(Attribute newAttribute) {
-		attribute = newAttribute;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	@Override
-	public String toString() {
-		StringBuffer result = new StringBuffer(super.toString());
-		result.append(" (instant: ");
-		result.append(instant);
-		result.append(", value: ");
-		result.append(value);
-		result.append(')');
-		return result.toString();
-	}
-
-} //HistoricalDataImpl
