--- analyzer/src/main/java/eubrazil/atmosphere/qualitymodel/impl/LeafAttributeImpl.java
@@ -1,192 +1,0 @@
-/**
- */
-package eubrazil.atmosphere.qualitymodel.impl;
-
-import eubrazil.atmosphere.qualitymodel.LeafAttribute;
-import eubrazil.atmosphere.qualitymodel.Metric;
-import eubrazil.atmosphere.qualitymodel.MetricAggregationOperator;
-
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * <!-- begin-user-doc -->
- * An implementation of the model object '<em><b>Leaf Attribute</b></em>'.
- * <!-- end-user-doc -->
- * <p>
- * The following features are implemented:
- * </p>
- * <ul>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.impl.LeafAttributeImpl#getMin <em>Min</em>}</li>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.impl.LeafAttributeImpl#getMax <em>Max</em>}</li>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.impl.LeafAttributeImpl#getMetric <em>Metric</em>}</li>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.impl.LeafAttributeImpl#getOperator <em>Operator</em>}</li>
- * </ul>
- *
- * @generated
- */
-public class LeafAttributeImpl extends AttributeImpl implements LeafAttribute {
-	/**
-	 * The default value of the '{@link #getMin() <em>Min</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getMin()
-	 * @generated
-	 * @ordered
-	 */
-	protected static final double MIN_EDEFAULT = 0.0;
-
-	/**
-	 * The cached value of the '{@link #getMin() <em>Min</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getMin()
-	 * @generated
-	 * @ordered
-	 */
-	protected double min = MIN_EDEFAULT;
-
-	/**
-	 * The default value of the '{@link #getMax() <em>Max</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getMax()
-	 * @generated
-	 * @ordered
-	 */
-	protected static final double MAX_EDEFAULT = 1.0;
-
-	/**
-	 * The cached value of the '{@link #getMax() <em>Max</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getMax()
-	 * @generated
-	 * @ordered
-	 */
-	protected double max = MAX_EDEFAULT;
-
-	/**
-	 * The cached value of the '{@link #getMetric() <em>Metric</em>}' reference list.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getMetric()
-	 * @generated
-	 * @ordered
-	 */
-	protected List<Metric> metric;
-
-	/**
-	 * The default value of the '{@link #getOperator() <em>Operator</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getOperator()
-	 * @generated
-	 * @ordered
-	 */
-	protected static final MetricAggregationOperator OPERATOR_EDEFAULT = MetricAggregationOperator.AVERAGE;
-
-	/**
-	 * The cached value of the '{@link #getOperator() <em>Operator</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getOperator()
-	 * @generated
-	 * @ordered
-	 */
-	protected MetricAggregationOperator operator = OPERATOR_EDEFAULT;
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	protected LeafAttributeImpl() {
-		super();
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public double getMin() {
-		return min;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public void setMin(double newMin) {
-		min = newMin;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public double getMax() {
-		return max;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public void setMax(double newMax) {
-		max = newMax;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public List<Metric> getMetric() {
-		if (metric == null) {
-			metric = new ArrayList<Metric>();
-		}
-		return metric;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public MetricAggregationOperator getOperator() {
-		return operator;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public void setOperator(MetricAggregationOperator newOperator) {
-		operator = newOperator == null ? OPERATOR_EDEFAULT : newOperator;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	@Override
-	public String toString() {
-		StringBuffer result = new StringBuffer(super.toString());
-		result.append(" (min: ");
-		result.append(min);
-		result.append(", max: ");
-		result.append(max);
-		result.append(", operator: ");
-		result.append(operator);
-		result.append(')');
-		return result.toString();
-	}
-
-} //LeafAttributeImpl
