--- analyzer/src/main/java/eubrazil/atmosphere/qualitymodel/impl/MetricImpl.java
@@ -1,164 +1,0 @@
-/**
- */
-package eubrazil.atmosphere.qualitymodel.impl;
-
-import eubrazil.atmosphere.qualitymodel.Metric;
-
-/**
- * <!-- begin-user-doc -->
- * An implementation of the model object '<em><b>Metric</b></em>'.
- * <!-- end-user-doc -->
- * <p>
- * The following features are implemented:
- * </p>
- * <ul>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.impl.MetricImpl#getProbe <em>Probe</em>}</li>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.impl.MetricImpl#getDescription <em>Description</em>}</li>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.impl.MetricImpl#getResource <em>Resource</em>}</li>
- * </ul>
- *
- * @generated
- */
-public class MetricImpl implements Metric {
-	/**
-	 * The default value of the '{@link #getProbe() <em>Probe</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getProbe()
-	 * @generated
-	 * @ordered
-	 */
-	protected static final int PROBE_EDEFAULT = 0;
-
-	/**
-	 * The cached value of the '{@link #getProbe() <em>Probe</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getProbe()
-	 * @generated
-	 * @ordered
-	 */
-	protected int probe = PROBE_EDEFAULT;
-
-	/**
-	 * The default value of the '{@link #getDescription() <em>Description</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getDescription()
-	 * @generated
-	 * @ordered
-	 */
-	protected static final String DESCRIPTION_EDEFAULT = null;
-
-	/**
-	 * The cached value of the '{@link #getDescription() <em>Description</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getDescription()
-	 * @generated
-	 * @ordered
-	 */
-	protected String description = DESCRIPTION_EDEFAULT;
-
-	/**
-	 * The default value of the '{@link #getResource() <em>Resource</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getResource()
-	 * @generated
-	 * @ordered
-	 */
-	protected static final int RESOURCE_EDEFAULT = 0;
-
-	/**
-	 * The cached value of the '{@link #getResource() <em>Resource</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getResource()
-	 * @generated
-	 * @ordered
-	 */
-	protected int resource = RESOURCE_EDEFAULT;
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	protected MetricImpl() {
-		super();
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public int getProbe() {
-		return probe;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public void setProbe(int newProbe) {
-		probe = newProbe;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public String getDescription() {
-		return description;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public void setDescription(String newDescription) {
-		description = newDescription;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public int getResource() {
-		return resource;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public void setResource(int newResource) {
-		resource = newResource;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	@Override
-	public String toString() {
-		StringBuffer result = new StringBuffer(super.toString());
-		result.append(" (probe: ");
-		result.append(probe);
-		result.append(", description: ");
-		result.append(description);
-		result.append(", resource: ");
-		result.append(resource);
-		result.append(')');
-		return result.toString();
-	}
-
-} //MetricImpl
