--- analyzer/src/main/java/eubrazil/atmosphere/qualitymodel/impl/UserProfileImpl.java
@@ -1,123 +1,0 @@
-/**
- */
-package eubrazil.atmosphere.qualitymodel.impl;
-
-import eubrazil.atmosphere.qualitymodel.Preference;
-import eubrazil.atmosphere.qualitymodel.UserProfile;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
-import org.eclipse.emf.common.notify.Notification;
-import org.eclipse.emf.common.notify.NotificationChain;
-
-import org.eclipse.emf.common.util.EList;
-
-import org.eclipse.emf.ecore.EClass;
-import org.eclipse.emf.ecore.InternalEObject;
-
-import org.eclipse.emf.ecore.impl.ENotificationImpl;
-
-import org.eclipse.emf.ecore.util.EObjectContainmentEList;
-import org.eclipse.emf.ecore.util.InternalEList;
-
-/**
- * <!-- begin-user-doc -->
- * An implementation of the model object '<em><b>User Profile</b></em>'.
- * <!-- end-user-doc -->
- * <p>
- * The following features are implemented:
- * </p>
- * <ul>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.impl.UserProfileImpl#getPreference <em>Preference</em>}</li>
- *   <li>{@link eubrazil.atmosphere.qualitymodel.impl.UserProfileImpl#getUserID <em>User ID</em>}</li>
- * </ul>
- *
- * @generated
- */
-public class UserProfileImpl implements UserProfile {
-	/**
-	 * The cached value of the '{@link #getPreference() <em>Preference</em>}' containment reference list.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getPreference()
-	 * @generated
-	 * @ordered
-	 */
-	protected List<Preference> preference;
-
-	/**
-	 * The default value of the '{@link #getUserID() <em>User ID</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getUserID()
-	 * @generated
-	 * @ordered
-	 */
-	protected static final int USER_ID_EDEFAULT = 0;
-
-	/**
-	 * The cached value of the '{@link #getUserID() <em>User ID</em>}' attribute.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @see #getUserID()
-	 * @generated
-	 * @ordered
-	 */
-	protected int userID = USER_ID_EDEFAULT;
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	protected UserProfileImpl() {
-		super();
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public List<Preference> getPreference() {
-		if (preference == null) {
-			preference = new ArrayList<Preference>();
-		}
-		return preference;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public int getUserID() {
-		return userID;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public void setUserID(int newUserID) {
-		userID = newUserID;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	@Override
-	public String toString() {
-		StringBuffer result = new StringBuffer(super.toString());
-		result.append(" (userID: ");
-		result.append(userID);
-		result.append(')');
-		return result.toString();
-	}
-
-} //UserProfileImpl
